Finishing fitseq

+++ src/io/image_format_fits.c
@@ -2153,18 +2153,8 @@ int siril_fits_compress(fits *f) {
 	return status;
 }
 
-/* creates, saves and closes the file associated to f, overwriting previous  */
-int savefits(const char *name, fits *f) {
-	int status;
-	char filename[256];
-
-	f->naxes[0] = f->rx;
-	f->naxes[1] = f->ry;
-
-	if (f->naxis == 3 && f->naxes[2] != 3) {
-		printf("Trying to save a FITS color file with more than 3 channels?");
-		return 1;
-	}
+gchar *set_right_extension(const char *name) {
+	gchar *filename = NULL;
 
 	gboolean comp_flag = FALSE;
 	/* first check if there is fz extension */
@@ -2190,29 +2180,49 @@ int savefits(const char *name, fits *f) {
 
 	if (!right_extension) {
 		if (com.pref.comp.fits_enabled) {
-			snprintf(filename, 255, "%s%s.fz", name, com.pref.ext);
+			filename = g_strdup_printf("%s%s.fz", name, com.pref.ext);
 		} else {
-			snprintf(filename, 255, "%s%s", name, com.pref.ext);
+			filename = g_strdup_printf("%s%s", name, com.pref.ext);
 		}
 	} else {
 		if (comp_flag && !com.pref.comp.fits_enabled) {
 			/* we remove .fz */
 			gchar *tmp = g_strdup(name);
 			tmp[strlen(tmp) - 3] = '\0';
-			snprintf(filename, 255, "%s", tmp);
+			filename = g_strdup_printf("%s", tmp);
+
 			g_free(tmp);
 		} else if (!comp_flag && com.pref.comp.fits_enabled) {
-			snprintf(filename, 255, "%s.fz", name);
+			filename = g_strdup_printf("%s.fz", name);
+
 		} else {
-			snprintf(filename, 255, "%s", name);
+			filename = g_strdup_printf("%s", name);
 		}
 	}
+	return filename;
+}
+
+/* creates, saves and closes the file associated to f, overwriting previous  */
+int savefits(const char *name, fits *f) {
+	int status;
+
+	f->naxes[0] = f->rx;
+	f->naxes[1] = f->ry;
+
+	if (f->naxis == 3 && f->naxes[2] != 3) {
+		printf("Trying to save a FITS color file with more than 3 channels?");
+		return 1;
+	}
+
+	gchar *filename = set_right_extension(name);
+	if (!filename) return 1;
 
 	g_unlink(filename); /* Delete old file if it already exists */
 
 	status = 0;
 	if (siril_fits_create_diskfile(&(f->fptr), filename, &status)) { /* create new FITS file */
 		report_fits_error(status);
+		g_free(filename);
 		return 1;
 	}
 
@@ -2220,12 +2230,14 @@ int savefits(const char *name, fits *f) {
 		status = siril_fits_compress(f);
 		if (status) {
 			report_fits_error(status);
+			g_free(filename);
 			return 1;
 		}
 	}
 
 	if (fits_create_img(f->fptr, f->bitpix, f->naxis, f->naxes, &status)) {
 		report_fits_error(status);
+		g_free(filename);
 		return 1;
 	}
 
@@ -2238,6 +2250,7 @@ int savefits(const char *name, fits *f) {
 				filename, f->naxes[2], f->rx, f->ry,
 				f->type == DATA_USHORT ? 16 : 32);
 	}
+	g_free(filename);
 	return 0;
 }
